refactor(backend): move stream-parsing utils to modules/claude (phase 3 step 1)
