+++ backend/modules/claude/content-block-expander.util.ts
