+++ backend/modules/claude/line-buffer.util.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach } from 'vitest';
-import { LineBuffer } from './lineBuffer.js';
+import { LineBuffer } from './line-buffer.util.js';
 
 describe('LineBuffer', () => {
   let buf: LineBuffer;
