+++ backend/modules/claude/line-buffer.util.ts
