+++ backend/modules/claude/stream-json-parser.util.ts
