+++ backend/services/claude/ClaudeProcess.ts
@@ -1,8 +1,8 @@
 import { spawn, type ChildProcess } from 'node:child_process';
 import { EventEmitter } from 'node:events';
 import { randomUUID } from 'node:crypto';
-import { LineBuffer } from './lineBuffer.js';
-import { parseStreamJsonLine } from './streamJsonParser.js';
+import { LineBuffer } from '../../modules/claude/line-buffer.util.js';
+import { parseStreamJsonLine } from '../../modules/claude/stream-json-parser.util.js';
 import { ClaudeArgsBuilder } from './ClaudeArgsBuilder.js';
 import type {
   ClaudeExecuteOptions,
